CIMCORE/canonical json updates (#19)
CIMCORE is generated through a strict definition of JSON output in each of the model files. It was done this way (as opposed to creating a separate CIMCORE output tool) in order to limit an unnecessary increase in tools and dependencies.

Between the various components of `DataElement`, `ValueSets`, and `Mapping`, the `toJSON` methods are meant to be compositional, building off each other and their parents.

The majority of the 'logic' can be found in the `toJSON` function of `Value`. That is where the more complex hierarchical constraint structure is found.

This is a SQUASHED commit of the following original commits:

* Initial Canonical Json

This is the barebones of the canonical json model, specifically the data elements component.

* Hierarchial Constraints

This adds hierarchy and history information to the constraints.

* ValueSet json

This is the preliminary output for the value sets.

* Mappings

Preliminary Mappings output.

* Updates based on review meeting suggestions

Additionally:
Added clearEmptyObject, and changed value/field structure

* Updates on readibility

Additionally added some undefined logic

* Added "lastmodifiedBy"

Additionally added getters and setters for "history" and reorganized Json output

* Unit tests

* General Updates and bug fixes

changes 'valueset' to 'uri'

fixes bug where subpath constraints would get duplicated

* Cardinality fix to handle new effective card

This was a merge issue when building on top of master. the cimcore was built on top of an old effective card.

* Package bump

* Added inheritedFrom convience Function

This is necessary to make unit tests easier.

* Various PR Review changes

* Fixed tests

* Inherited From Minor Updates

* Simplifications, convenience functions, and cleanups for LastModifiedBy, Cardinality History, and InheritedFrom

(Related to large unit test updates in shr-expand)

Simplifications:
- simplified lastModifiedBy to be an ‘identifier’ (as opposed to a
pointer to full data element)
-  simplified InheritedFrom to be an ‘identifier’ (as opposed to a
pointer to full data element)
- simplified source to be an ‘identifier’ (as opposed to a pointer to
full data element)

Convenience functions:
- added source and history convenience functions to Cardinality (to
assist in unit test fixes)
- added LastModifiedBy convenience function to Constraints (to assist
in unit test fixes)

Cleanups:
- removed all ‘Foo = undefined’ and ‘Bar = []’ defaults from
constructor (by improving toJSON output logic)
- some comments

* Defensive check and comment update