Must Support (#86)
* WIP add content profile grammar and begin listener

* Add data to content profile specifications (except path)

* Add content profile listener and add config file field for content profile

* Remove extraneous comment

* Add handling for invalid content profile file in config

* Add support for Value in chain, fix styling, add extra error handling

* Add null checks for configuration

* Clear currentDef and currentRule when appropriate

When the parser exits the CP definition or a CP rule definition, clear the corresponding member property.  Otherwise old rules can get applied to new CPs!

* Add support for inherited attributes in content profiles

* Address comments based on review feedback

* Content Profile paths must use the effectiveIdentifier

After discussion w/ Mark, we determined that CPs should always use the "effective identifiers" in the paths.  For example, if field Foo is constrained to SonOfFoo, then the CP path should use SonOfFoo.

In addition fixed an issue where the ChoiceValue options were being iterated as identifiers when they are really Values (that have identifiers).

* Minor tweak for clarity (based on code review)

* Update error numbers

* Upgrade shr-models to 5.8.0

Take in the new models for ContentProfiles.

* v5.7.0