Make title element more human-readable
The title attributes on structure definitions, extensions, and logical models make use of
a new "title" function on the Identifier class. This function uses the Identifier's name and
inserts spaces based on word boundary conditions. Because a title may contain letters
and numbers, it is necessary to use somewhat unusual-looking statements when
checking for word boundaries.