Additional use of Identifier title in FHIR and Data Dictionary exports
In FHIR export, use Identifier title in code systems and value sets.
In Data Dictionary export, use Identifier title for profile and data element names.
Note that the Data Dictionary does not use the Identifier title for Value Sets.
The regular expressions previously used to create the title are removed.