Report error when element added without Property keyword
New properties must use the Property keyword when they are
added. It is permissible to not use the Property keyword
when constraining a property that has already been declared,
either on the current element or an ancestor of the element.

As an example, this error helps in detecting cases where an
author intended to constrain an inherited property, but did
not spell the name of the property correctly.

Additionally, an error logged when a property is redundantly
declared now logs correctly.