Clarify text on Code Systems IG page
The text explaining what code systems are listed on the IG's Code Systems
page is clarified in regards to what exactly is listed and why. The explanation
varies depending on whether or not the page only lists primary code systems.

If there are no code systems to include in the IG, delete the Code Systems
page and hide the navbar's link to the Code Systems page.