Do not delete codesystems.html
Since codesystems.html is reference in some of the templates (like nav) and breadcrumbs, the IG publisher complains if the file is missing (even if it's not exposed).  This change reverts a previous change that deleted the codesystems.html file when there were no code systems.  It also refactors some of the text back to the html template.