Account for "compiler magic" in 3.0.2 and 4.0.1
In the 3.0.2 and 4.0.1 releases, Grahame added what he calls "compiler magic" -- which causes some primitive types to have a type code that is a FHIRPath URL.  This mainly should not affect us except when someone maps something to a resource's `id` property -- but sure enough, someone is doing that. ;-) . This resolves it by allowing a conversion from primitive types to their corresponding FHIRPath URLs.